+++ library/src/androidTest/java/de/nenick/espressomacchiato/elements/EspViewTest.java
@@ -24,6 +24,8 @@ public class EspViewTest extends EspressoTestCase<BaseActivity> {
 
     @Test
     public void testAssertions() {
+        espTextView.assertNotExist();
+
         givenClickableView();
 
         espView.assertIsVisible();