+++ RELEASE_NOTES.md
@@ -4,9 +4,6 @@ planed
 
 * EspPermissions - accept/deny dialog with ui automator
 * EspCameraMock - https://github.com/googlesamples/android-testing/tree/master/ui/espresso/IntentsBasicSample
-* EspSimpleOneButtonDialog
-* EspSimpleTwoButtonsDialog
-* EspSimpleThreeButtonsDialog
 * EspSimpleButtonListDialog
 
 done
@@ -14,6 +11,11 @@ done
 * All - extend find view strategy
 * All - static factory methods to hide implementation details for changes
 * EspView - click only visible views
+* EspView - does not exist in view hierarchy assertion
 * EspDevice - close soft keyboard and do assertions for it
+* EspDialog - new element
+* EspOneButtonDialog - new element
+* EspTwoButtonsDialog - new element
+* EspThreeButtonsDialog - new element
 * Dependencies - don't exclude junit, just ignore lint error
 * Dependencies - update espresso 2.2.1 to 2.2.2